+++ Spells/Spell.cpp
@@ -164,27 +164,27 @@ void Spell::increment_full_block() {
     statistics_spell->increment_full_block();
 }
 
-void Spell::add_partial_resist_dmg(const int damage, const unsigned resource_cost, const double execution_time) {
+void Spell::add_partial_resist_dmg(const int damage, const double resource_cost, const double execution_time) {
     statistics_spell->add_partial_resist_dmg(damage, resource_cost, execution_time);
 }
 
-void Spell::add_partial_block_dmg(const int damage, const unsigned resource_cost, const double execution_time) {
+void Spell::add_partial_block_dmg(const int damage, const double resource_cost, const double execution_time) {
     statistics_spell->add_partial_block_dmg(damage, resource_cost, execution_time);
 }
 
-void Spell::add_partial_block_crit_dmg(const int damage, const unsigned resource_cost, const double execution_time) {
+void Spell::add_partial_block_crit_dmg(const int damage, const double resource_cost, const double execution_time) {
     statistics_spell->add_partial_block_crit_dmg(damage, resource_cost, execution_time);
 }
 
-void Spell::add_glancing_dmg(const int damage, const unsigned resource_cost, const double execution_time) {
+void Spell::add_glancing_dmg(const int damage, const double resource_cost, const double execution_time) {
     statistics_spell->add_glancing_dmg(damage, resource_cost, execution_time);
 }
 
-void Spell::add_hit_dmg(const int damage, const unsigned resource_cost, const double execution_time) {
+void Spell::add_hit_dmg(const int damage, const double resource_cost, const double execution_time) {
     statistics_spell->add_hit_dmg(damage, resource_cost, execution_time);
 }
 
-void Spell::add_crit_dmg(const int damage, const unsigned resource_cost, const double execution_time) {
+void Spell::add_crit_dmg(const int damage, const double resource_cost, const double execution_time) {
     statistics_spell->add_crit_dmg(damage, resource_cost, execution_time);
 }
